Add RETURN_INTO_ANCESTOR attack
This attack calls a function that has been called previously
and could be found in a stack trace, i.e., it is an ancestor
of the attacked function.

Some CFI schemes try to limit returns to such functions that
have been visited in the CFG to this point (without returning).
This attack would not be detected by these schemes.