arm: implement longjmp from exception handlers
Returning from exceptions is peculiar in ARM as there
are no special instructions to end exception mode.
Instead the code has to load a special value into
the PC register that is passed to the handler in LR
by hardware.

This adds an assembly routine to "call" longjmp_no_enforce()
from an exception and use it for all kind of exceptions
instead of endless loops.